Send EndOfRib marker to xmpp agents after sweep of all tables are complete
During GR, if the sessions come back up, then we exit out of GR either because
agent sends EndOfRib marker or because the timer expires. In either case, we
sweep all the tables. After this process is complete, we can send an EndOfRib
marker to the agent so that agent can sweep its tables as well

Change-Id: Iaaab2572510209737ff23a7b7d34e2dc69882efc
Partial-Bug: #1537933