Spread receiver and sender across different virtual-networks
Sender is connected to one control-node and receiver is connected
to another. Hence the routes are learned over bgp in this scenario.

Fix a couple of issues found there by. In particular, when Type-4 path
needs to be replicated, we can identify the target table by extracting
the type-3 prefix from the type-4 prefix, and then get to the target
table using the src_table() of the associated imported secondary
type-3 path.

Partial-Bug: #1708403